pyproject.toml
[project]
name = "langchain-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
]
